+++ src/main/java/com/sumo/server/Database/userData/PersonalDetails/PersonalDetails.java
@@ -1,0 +1,46 @@
+package com.sumo.server.Database.userData.PersonalDetails;
+
+import com.sumo.server.Database.StaticData.Sex.Sex;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import javax.persistence.*;
+import java.time.LocalDate;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Table(name = "PERSONAL_DETAILS")
+public class PersonalDetails {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.SEQUENCE)
+    private Long id;
+
+    @Column(name = "NAME")
+    private String name;
+
+    @Column(name = "SURNAME")
+    private String surname;
+
+    @Column(name = "PHONE_NUMBER")
+    private String phoneNumber;
+
+    @Column(name = "EMAIL")
+    private String email;
+
+    @Column(name = "LINK_TO_PROFILE_PICTURE")
+    private String linkToProfilePicture;
+
+    @Column(name = "BIRTH_DATE")
+    private LocalDate birthDate;
+
+    @ManyToOne
+    @JoinColumn(name = "SEX")
+    private Sex sex;
+}
