+++ src/main/java/com/sumo/server/Database/userData/User/UserService.java
@@ -1,7 +1,7 @@
-package com.sumo.server.userData.User;
+package com.sumo.server.Database.userData.User;
 
-import com.sumo.server.userData.PersonalDetails.PersonalDetails;
-import com.sumo.server.userData.Role.Role;
+import com.sumo.server.Database.userData.PersonalDetails.PersonalDetails;
+import com.sumo.server.Database.userData.Role.Role;
 
 import java.util.List;
