+++ src/main/java/com/sumo/server/apis/TokenService.java
@@ -6,8 +6,8 @@ import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.interfaces.DecodedJWT;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sumo.server.Seciurity.TokenGenerator;
-import com.sumo.server.userData.User.User;
-import com.sumo.server.userData.User.UserService;
+import com.sumo.server.Database.userData.User.User;
+import com.sumo.server.Database.userData.User.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.GetMapping;
